chore(modules): Rename error module to errors
